requirements.txt
Django==5.2.4
psycopg2-binary==2.9.10
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
python-decouple==3.8
django-environ==0.12.0
PyJWT==2.9.0
pytz==2025.2
tzdata==2025.2
sqlparse==0.5.3 